# service/service/employee/_swagger/service/swagger.yaml
openapi: 3.0.3
info:
  version: "1.0"
  title: "Employee Management"
  contact:
    email: "m.ehab.rabea@gmail.com"
    name: "Mohamed Ehab"
servers:
  - url: http://localhost:7002/
...
paths:
  /lookup/employee/attendance/status:
    get:
      tags:
        - Lookup
      summary: Retrieve All Employee Attendance Status (public)
      operationId: getAllEmployeeAttendanceStatusLookup
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/LookupResultSet'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
  /lookup/employee/types:
    get:
      tags:
        - Lookup
      summary: Retrieve All Employee Types (public)
      operationId: getAllEmployeeTypesLookup
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/LookupResultSet'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
  /lookup/employees:
    get:
      tags:
        - Lookup
      summary: Retrieve All Employees (public)
      operationId: getAllEmployeesLookup
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/LookupResultSet'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
...
  /trainers/{trainerId}/courses:
    get:
      tags:
        - TrainerCourse
      summary: Get all courses assigned to a trainer
      operationId: getTrainerCourses
      parameters:
        - $ref: '#/components/parameters/PPTrainerId'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageNum'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageSize'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderDir'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderBy'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerCourseResultSet'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'

    post:
      tags:
        - TrainerCourse
      summary: Assign a course to a trainer
      operationId: assignCourseToTrainer
      parameters:
        - $ref: '#/components/parameters/PPTrainerId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignCourseDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/NewRecordVTO'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
  /trainers/{trainerId}/courses/{courseId}:
    delete:
      tags:
        - TrainerCourse
      summary: Unassign a course from a trainer
      operationId: unassignCourseFromTrainer
      parameters:
        - $ref: '#/components/parameters/PPTrainerId'
        - $ref: '#/components/parameters/PPCourseId'
      responses:
        '200':
          description: OK
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
...
  /courses/{courseId}/sessions:
    post:
      tags:
        - CourseSession
      summary: Create course session
      operationId: createCourseSession
      parameters:
        - $ref: '#/components/parameters/PPCourseId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseSessionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/NewRecordVTO'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'

    get:
      tags:
        - CourseSession
      summary: Get all course sessions by filter
      operationId: getAllCourseSessionsByFilter
      parameters:
        - $ref: '#/components/parameters/PPCourseId'
        - $ref: '#/components/parameters/QPSessionStatus'
        - $ref: '#/components/parameters/QPSessionDateFrom'
        - $ref: '#/components/parameters/QPSessionDateTo'
        - $ref: '#/components/parameters/QPStartTimeFrom'
        - $ref: '#/components/parameters/QPStartTimeTo'
        - $ref: '#/components/parameters/QPEndTimeFrom'
        - $ref: '#/components/parameters/QPEndTimeTo'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageNum'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageSize'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderDir'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderBy'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseSessionResultSet'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
  /courses/sessions:
    get:
      tags:
        - CourseSession
      summary: Get all course sessions across all courses with filters
      operationId: getAllSessionsByFilter
      parameters:
        - $ref: '#/components/parameters/QPCourseId'
        - $ref: '#/components/parameters/QPTrainerId'
        - $ref: '#/components/parameters/QPPlaceId'
        - $ref: '#/components/parameters/QPSessionStatus'
        - $ref: '#/components/parameters/QPSessionDateFrom'
        - $ref: '#/components/parameters/QPSessionDateTo'
        - $ref: '#/components/parameters/QPStartTimeFrom'
        - $ref: '#/components/parameters/QPStartTimeTo'
        - $ref: '#/components/parameters/QPEndTimeFrom'
        - $ref: '#/components/parameters/QPEndTimeTo'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageNum'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPPageSize'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderDir'
        - $ref: '../../../../../lib/sql-db-adapter/_swagger/service/swagger.yaml#/components/parameters/QPOrderBy'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseSessionResultSet'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
  /courses/{courseId}/sessions/{sessionId}:
    put:
      tags:
        - CourseSession
      summary: Update course session
      operationId: updateCourseSession
      parameters:
        - $ref: '#/components/parameters/PPCourseId'
        - $ref: '#/components/parameters/PPSessionId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseSessionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/NewRecordVTO'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'

    get:
      tags:
        - CourseSession
      summary: Get course session by id
      operationId: getCourseSessionById
      parameters:
        - $ref: '#/components/parameters/PPCourseId'
        - $ref: '#/components/parameters/PPSessionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseSessionVTO'
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'

    delete:
      tags:
        - CourseSession
      summary: Delete course session (soft delete)
      operationId: deleteCourseSession
      parameters:
        - $ref: '#/components/parameters/PPCourseId'
        - $ref: '#/components/parameters/PPSessionId'
      responses:
        '200':
          description: OK
        '400':
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/responses/BadRequest'
components:
  parameters:
...
    PPTrainerId:
      name: trainerId
      in: path
      required: true
      schema:
        type: integer
        format: int32
    PPCourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: integer
        format: int32
    PPSessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: integer
        format: int32
...
    QPTrainerId:
      name: trainerId
      in: query
      required: false
      schema:
        type: integer
        format: int32
    QPCourseId:
      name: courseId
      in: query
      required: false
      schema:
        type: integer
        format: int32
...
    QPPlaceId:
      name: placeId
      in: query
      required: false
      schema:
        type: integer
        format: int32
    QPSessionStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/SessionStatus'
    QPSessionDateFrom:
      name: sessionDateFrom
      in: query
      required: false
      schema:
        type: string
        format: date
    QPSessionDateTo:
      name: sessionDateTo
      in: query
      required: false
      schema:
        type: string
        format: date
    QPStartTimeFrom:
      name: startTimeFrom
      in: query
      required: false
      schema:
        type: string
        format: time
    QPStartTimeTo:
      name: startTimeTo
      in: query
      required: false
      schema:
        type: string
        format: time
    QPEndTimeFrom:
      name: endTimeFrom
      in: query
      required: false
      schema:
        type: string
        format: time
    QPEndTimeTo:
      name: endTimeTo
      in: query
      required: false
      schema:
        type: string
        format: time
  schemas:
...
    SessionStatus:
      type: object
      properties:
        id:
          type: integer
          enum: [ 1, 2, 3, 4 ]
        code:
          type: string
          enum: [ "SCHEDULED", "IN_PROGRESS", "COMPLETED", "CANCELLED" ]
        arabicLabel:
          type: string
          enum: [ "مجدول", "في تقدم", "مكتمل", "ملغي" ]
...
    CourseSessionDTO:
      type: object
      required:
        - title
        - trainerId
        - courseId
        - placeId
        - sessionDate
        - startTime
        - endTime
      properties:
        title:
          type: string
        courseId:
          type: integer
          format: int32
        trainerId:
          type: integer
          format: int32
        placeId:
          type: integer
          format: int32
        sessionDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        status:
          $ref: '#/components/schemas/SessionStatus'
        note:
          type: string
    AssignCourseDTO:
      type: object
      required:
        - courseId
      properties:
        courseId:
          type: integer
          format: int32
...
    TrainerCourseVTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        course:
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/LookupVTO'
        createdOn:
          type: string
          format: date-time
        createdBy:
          $ref: '../../../../../service/service/user/_swagger/service/swagger.yaml#/components/schemas/LightUserVTO'
...
    TrainerCourseResultSet:
      type: object
      properties:
        total:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/TrainerCourseVTO'
...
    CourseSessionVTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        course:
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/LookupVTO'
        trainer:
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/LookupVTO'
        place:
          $ref: '../../../../../lib/common/_swagger/service/swagger.yaml#/components/schemas/LookupVTO'
        sessionDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        status:
          $ref: '#/components/schemas/SessionStatus'
        note:
          type: string
        createdOn:
          type: string
          format: date-time
        createdBy:
          $ref: '../../../../../service/service/user/_swagger/service/swagger.yaml#/components/schemas/LightUserVTO'
        lastModifiedOn:
          type: string
          format: date-time
        lastModifiedBy:
          $ref: '../../../../../service/service/user/_swagger/service/swagger.yaml#/components/schemas/LightUserVTO'
    CourseSessionResultSet:
      type: object
      properties:
        total:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/CourseSessionVTO'